+++ src/ts/anomalies/mandelbulb/mandelbulbModelGenerator.ts
@@ -24,8 +24,14 @@ import { getRngFromSeed } from "../../utils/getRngFromSeed";
 import { MandelbulbModel } from "./mandelbulbModel";
 import { Color3 } from "@babylonjs/core/Maths/math.color";
 import { Tools } from "@babylonjs/core/Misc/tools";
+import { OrbitalObjectModel } from "../../architecture/orbitalObjectModel";
+import { createOrbitalObjectId } from "../../utils/coordinates/orbitalObjectId";
 
-export function newSeededMandelbulbModel(seed: number, name: string): MandelbulbModel {
+export function newSeededMandelbulbModel(
+    seed: number,
+    name: string,
+    parentBodies: ReadonlyArray<OrbitalObjectModel>
+): MandelbulbModel {
     const rng = getRngFromSeed(seed);
 
     const radius = 1000e3;
@@ -42,7 +48,10 @@ export function newSeededMandelbulbModel(seed: number, name: string): Mandelbulb
 
     const orbitalP = clamp(0.5, 3.0, normalRandom(1.0, 0.3, rng, GenerationSteps.ORBIT + 80));
 
+    const parentIds = parentBodies.map((body) => body.id);
+
     const orbit: Orbit = {
+        parentIds: parentIds,
         semiMajorAxis: orbitRadius,
         p: orbitalP,
         inclination: Tools.ToRadians(normalRandom(90, 20, rng, GenerationSteps.ORBIT + 160)),
@@ -56,9 +65,10 @@ export function newSeededMandelbulbModel(seed: number, name: string): Mandelbulb
     const axialTilt = normalRandom(0, 0.4, rng, GenerationSteps.AXIAL_TILT);
 
     return {
-        radius,
-        name,
         type: OrbitalObjectType.MANDELBULB,
+        id: createOrbitalObjectId(parentIds, name),
+        name,
+        radius,
         mass,
         siderealDaySeconds,
         axialTilt,
