+++ src/ts/tutorials/stationLandingTutorial.spec.ts
@@ -18,7 +18,7 @@
 import { describe, expect, it } from "vitest";
 import { StationLandingTutorial } from "./stationLandingTutorial";
 import { StarSystemDatabase } from "../starSystem/starSystemDatabase";
-import { getObjectModelByUniverseId } from "../utils/coordinates/orbitalObjectId";
+import { getObjectModelByUniverseId } from "../utils/coordinates/orbitalObjectIdUtils";
 import { OrbitalObjectType } from "../architecture/orbitalObjectType";
 
 describe("StationLandingTutorial", () => {
